Move cache related class into separate package
