+++ presto-hive/src/main/java/com/facebook/presto/hive/HiveClientModule.java
@@ -25,11 +25,11 @@ import com.facebook.presto.hive.rcfile.RcFilePageSourceFactory;
 import com.facebook.presto.hive.rule.HivePlanOptimizerProvider;
 import com.facebook.presto.hive.s3.PrestoS3ClientFactory;
 import com.facebook.presto.orc.CacheStatsMBean;
-import com.facebook.presto.orc.CachingOrcFileTailSource;
-import com.facebook.presto.orc.OrcCacheConfig;
+import com.facebook.presto.orc.cache.CachingOrcFileTailSource;
+import com.facebook.presto.orc.cache.OrcCacheConfig;
 import com.facebook.presto.orc.OrcDataSourceId;
-import com.facebook.presto.orc.OrcFileTailSource;
-import com.facebook.presto.orc.StorageOrcFileTailSource;
+import com.facebook.presto.orc.cache.OrcFileTailSource;
+import com.facebook.presto.orc.cache.StorageOrcFileTailSource;
 import com.facebook.presto.orc.metadata.OrcFileTail;
 import com.facebook.presto.spi.connector.ConnectorNodePartitioningProvider;
 import com.facebook.presto.spi.connector.ConnectorPageSinkProvider;
