+++ presto-orc/src/main/java/com/facebook/presto/orc/cache/StorageOrcFileTailSource.java
@@ -11,9 +11,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.orc;
+package com.facebook.presto.orc.cache;
 
 import com.facebook.airlift.log.Logger;
+import com.facebook.presto.orc.OrcCorruptionException;
+import com.facebook.presto.orc.OrcDataSource;
+import com.facebook.presto.orc.OrcWriteValidation;
 import com.facebook.presto.orc.metadata.CompressionKind;
 import com.facebook.presto.orc.metadata.MetadataReader;
 import com.facebook.presto.orc.metadata.OrcFileTail;
